Increase contrast/legibility of grid.
Changes:

- Draw a black background (with transparency) underneath the entire grid.
- Tweak initial alpha component for grid labels.

The downside of setting a solid background colour is that it makes it more
difficult to see the contents of the screen. (But, it does make it dramatically
easier to see the actual grid labels.)

Future improvement: dynamically tweak the transparency of the background on a
per-cell basis, based on the underlying content of the given cell.