Run as an “agent application”.
With this change, the Scoot app icon does not appear in the Dock, and it does
not have a menu bar.

(A benefit of not having a menu bar is that the app that was in the foreground
when Scoot is activated retains its menu bar; you can now use Scoot to click on
it.)

Note that, when Scoot is launched, the grid will no longer automatically
appear. You must invoke the global hotkey (⇧⌘J).

Reference material: https://stackoverflow.com/q/620841/15304124