Add "freestyle" navigation mode.
Freestyle is a freeform usage mode that offers no special
navigation assistance. (The existing support for moving the
cursor using document editing shortcuts, both MacOS + Emacs, will
work, as they do in the other modes.)

This mode is a stepping stone for enabling additional keyboard
shortcuts for moving the cursor. This includes vi shortcuts,
which would normally conflict with the sequences for navigating
the character-based decision trees of the grid-based and
element-based nav modes.